doublets::data

Trait DoubletsExt

pub trait DoubletsExt<T: LinkReference>: Sized + Doublets<T> {
    type IdxParIter: IndexedParallelIterator<Item = Link<T>>;

    // Required methods
    fn par_iter(&self) -> Self::IdxParIter;
    fn par_each_iter(&self, query: impl ToQuery<T>) -> Self::IdxParIter;
    fn iter(
        &self,
    ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator;
    fn each_iter(
        &self,
        query: impl ToQuery<T>,
    ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator;
    fn iter_small(
        &self,
    ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator;
    fn each_iter_small(
        &self,
        query: impl ToQuery<T>,
    ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator;
}

Extension trait that adds iterator-based access to a Doublets store.

Automatically implemented for any type that implements Doublets.

Required Associated Types

type IdxParIter: IndexedParallelIterator<Item = Link<T>>

The parallel iterator type returned by par_iter.

Required Methods

fn par_iter(&self) -> Self::IdxParIter

Returns a parallel iterator over all links in the store.

fn par_each_iter(&self, query: impl ToQuery<T>) -> Self::IdxParIter

Returns a parallel iterator over links matching query.

fn iter( &self, ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator

Returns an iterator over all links in the store.

fn each_iter( &self, query: impl ToQuery<T>, ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator

Returns an iterator over links matching query.

fn iter_small( &self, ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator

Returns a small-vec-backed iterator over all links (optimised for small result sets).

fn each_iter_small( &self, query: impl ToQuery<T>, ) -> impl Iterator<Item = Link<T>> + ExactSizeIterator + DoubleEndedIterator

Returns a small-vec-backed iterator over links matching query.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.